Renamed ap to ap_factory
